IECoreGL::Primitive : Fix wireframe rendering with user vertex shaders.
To perform wireframe rendering, we need to override the fragment shader so we can render a flat colour. Until now, we were also replacing the vertex shader with the default one, which meant the wireframe was in the wrong place if a vertex shader had been specified by the user.