IECoreGL : Fix IDRender selection with user vertex shaders.
Prior to this, the Selector base state contained an override shader which prevented any other shaders from being bound at all, so custom vertex shaders were not even visible to the Primitive in the render() method. Fixing this requires that the Selector base state has no such override, and it therefore becomes the responsibility of the Primitive to bind an appropriate ID shader when necessary. It does so using the same logic as for the constant wireframe shader, incorporating any user vertex shaders as necessary.