IECoreGL::Primitive : Fix selection with custom vertex shaders on OSX.
When we're falling back to simple rendering, we still need to take the vertex shader into account. We had originally used an even simpler shader than the constant to help with a very slow GL_SELECT with the NVIDIA drivers, but we shouldn't be following that code path any more except on OSX, where the constant shader seems fine.