LRUCache : Improve performance under heavy contention.
The previous version used a list to track the order of most recent use, moving an item to the end of the list each time it was accessed. This resulted in heavy contention for the mutex protecting the list, and poor performance. The new version uses an approximate strategy for tracking recent use, reducing contention in the performance critical case of an item already being in the cache.

This speeds up a call to `IECore.testLRUCacheThreading( 10000000, 100, 100, 0 )` by a factor of 10.