IECoreMaya: Preventing primary UV Set from being exported twice.
This hasn't been a problem in terms of disk space or memory because the s/t and map1_s/t data were indentified as identical and deduplicated, but it has been wasteful when we pass those extra primitive variables to some renderers. It was also preventing a round-trip of the data back into Maya in certain scenarios.