IECore python module : Use RTLD_GLOBAL to load C++ modules.
As the comment explains, this is necessary when several modules share the same dynamic library dependencies. We were sprinkling this workaround around liberally in various places using Cortex, so it makes sense to just do it once and for all in a central place, so clients can just `import IECore` and not need to worry.

In this instance, this is motivated by spending several hours tracking down a crash in the IECoreArnold tests, but it could apply anywhere multiple modules link to Cortex.