IECoreArnold::Renderer : Support "ai:shape:" attribute prefix.
Attributes with this prefix are added to all shapes, in the same fashion as attributes with the prefix "ai:nodeType:" are. This is primarily of use for setting the "step_size" for volume rendering, which is the same on boxes and spheres, and presumably will also be used on other supported types in the future.