Added mightHaveFn arg to registerCustomAttributes.
IECoreMaya::LiveScene::registerCustomAttributes() takes the third argument mightHaveFn.
If the mightHave function is specified, it will be called before names function for early out, to see if the names function can return the expected attribute.
It was added because for some custom attributes names function can be the performance bottle neck.