IECoreArnold::MeshAlgo : Suppress conversion of stIndices primvar.
This is sidecar data used to disambiguate the indexing of st, because the RenderMan specification we've used for FaceVarying data throws away valuable information. Perhaps at some point we'll need to use the indices to improve our UV conversion to Arnold, but we'll never want to convert the stIndices explicitly to user data on their own.