DisplayDriverServer : Support automatic selection of free port.
Passing an explicit port number is error prone; even when a free port is searched for prior to construction, race conditions can mean the port is no longer free when the constructor runs. We now encourage passing a port number of 0 to select a free port internally, and provide a `portNumber()` method to retrieve the port chosen.