Add a parameter to on/off checking the normal vector direction
Add a parameter to on/off checking if the normal vector direction is towards the cross product of uTangent and vTangent.
This check was already there for a while but we found a case where sudden change of uTangent vector can cause a problem during animation.
We can skip this check when uTangent flipping causes more problem than normal vector direction.