C++ tests : Fix crash on Travis.
We were using a deprecated mechanism for initialising the unit tests, and that appears to actually be broken with the Travis boost build. We now use a non-deprecated (but not particularly recommended) form instead, since this represents the least change required to fix the crash. In the long run we'd be better to do what we've done in Gaffer, and bind any C++ test functions so they can be integrated with the python test suite.