added an option to convert scene tags automatically to houdini groups
Added one new parameter on the scene cache source node - this will
turn creating groups from tagged geometry in scene caches on and off.
This will help save artists time and prevent errors from manually
creating groups.

Groups are created with the prefix "ieTag_". ObjectType:* tags will not
show up as groups. This is because the ieTag_ groups are recognized
by export procedures that will create tags from these groups, and
because ObjectType: tags are not meant to be added by accident.