added H16 required changes to the Houdini Cortex Primitive
This includes changes to the way that the primitives are
registered in the plugin, vertex allocation, and some
changes to which functions require implementation.

Vertex allocation no longer allowed in constructor. Since the
initial vertex offset isn't required to be stored anymore, I removed
some of the functions for H16 that would manage that offset in
the primitive.