ClientDisplayDriver : Recover from socket send errors
The `send()` method can be interrupted before, resulting in the wrong amount of data being sent, and the server reporting errors. This was most readily seen when using Ctrl+Z/bg in a shell to background a Gaffer session, or when running Gaffer in a VTune session. The `write()` method automatically recovers from such problems and sends the rest of the data.

In theory this could be slower because `write()` must wait before returning, but timings taken during interactive renders in Gaffer showed no difference. It may be worth testing this with a remote render over a network though.