IECoreArnold::ParameterAlgo : Fix string array conversion on OSX
The default conversion case was assuming that `std::string` is laid out in memory such that it is equivalent to `const char *`. That may have been the case in libstdc++, but it is not the case for libc++. I also suspect that it is not the case in libstdc++ with _GLIBCXX_USE_CXX11_ABI defined.