AlembicInputTest : Simplify and fix test for archive with no top level bound
The `boundAtSample( 0 )` call was throwing and causing a test failure. But it _should_ have been throwing, because the sample argument was out of range. I've removed the `objectAtSample()` and `transformAtSample()` calls because they were also throwing, but the AlembicInput API doesn't provide enough information to know that upfront - because it uses one sample count for everything. This will be resolved when we replace AlembicInput with AlembicScene.