LRUCache : Add Policy template argument
Policies implement the internal data structures for the cache, and different policies can be specialised for different circumstances. Initially there are two policies - a Serial policy which provides improved performance when thread safety is not required, and a Parallel policy which is slower but provides thread safety.

There's a lot happening in this commit, because neither policy shares a great deal with the previous implementation. When reviewing it might be easiest to read the documentation and comments for the Serial policy to get the general gist, and then go from there.