ImageWriter : Provide color conversion based on file extension.
By default, channel data will be converted from linear to the colorspace specified based on the file format and ImageSpec. This uses a naive approach, based on the current OIIO ColorConfig. If you built OpenImageIO with OpenColorIO support, this is the color manager used.

Note that some images may already be in the correct colorspace, if they were read directly from an ImageReader for example. If this is the case, use the rawChannels parameter to indicate we don't want any color conversion performed by the writer. Note that the formatSettings still take precedence for the data type.