IECoreGL : Read raw channels in various test images.
The Disk, Mesh, and Points tests were incorrectly changed in 6b494b9b916fe06ed4a523ed4549e8da97853c71, which had been informing the old ImageReader that the data was already linear.

TextTest had been failing for a long time, and adding the same "raw" setting to the reader allows it to pass, so presumably this was an accidental omission at some point in the past.