Build : Enable building IECoreHoudini without IECoreGL
This seems generally beneficial to others, but the main reason I'm doing it now is to temporarily skirt the versionitis issues casued by Houdini 16.0 not following the VFX Platform.

At Image Engine, I've disabled IECoreGL when building IECoreHoudini (see changes to config/ie/options). This is a temprorary workaround for the clash between OpenImageIO and Houdini dependencies. With GL enabled, libIECoreHoudini links to libIECoreGL links to libIECoreImage links to libOpenImageIO. At Image Engine, the only available libOpenImageIO suitabe for Cortex 10 links to Boost 1.61.0, whereas Houdini itself links to Boost 1.55.0, in blatant contradiction of VFX Platform 2017. By the time Cortex 10 is in a release state, Houdini 16.5 should be released, as well as several other OpenImageIO versions, so we should hopefully be able to re-enable it in time.