OpenImageIOAlgo : Refactor DataView
This was intended to be a non-owning view onto IECore::Data, analogous to std::string_view but for use when passing Data to OIIO. In 0a93e1314d58bb9b63d9054fb66b3ff1eba93b28 it acquired the opposite semantics, owning a DataPtr converted _from_ OIIO::ParamValue. This commit simply refactors the latter functionality out into a separate function, as it was before it was ported over from Gaffer.