MurmurHash : update data types to avoid crashes
We were running into an issue where the number of bytes passed in as a size_t was
2399213736, which is larger than what the signed ints used internally can
hold. Dividing by 16 and then multiplying by 16 didn't result in what we passed
in but in -1895753568. The address that tail points to is then not inside the
range of the given data causing a crash.