CurvesAlgo::resamplePrimitiveVariables() : Update for prim var indices.
Note that we force index expansion for downsampling, rather than recompute the indexed data in a smarter way.

Also note that we're accepting a less than ideal compromise here, where Vertex <--> Varying/FaceVarying throws if the input PrimitiveVariable has indices. This is because the functors to do the resampling are using CurvesPrimitiveEvaluators and requiring the resampled variable to be part of the input CurvesPrimitive, which is const, so we can't expand the indices.