ToGLMeshConverter : Don't flip V coordinate
Cortex has the following conventions :

- 0,0 in image pixel space is top left, with increasing Y going down
- 0,0 in UV space is bottom left, with increasing V going up

OpenGL has the following conventions :

- 0,0 in texture pixel space is bottom left, with increasing Y going up
- 0,0 in UV space is bottom left, with increasing V going up

I haven't found chapter and verse for this, but I believe
[this link](https://stackoverflow.com/questions/17430753/what-is-the-orientation-of-opengl-texture-data)
because it matches my own observations.

Since the image/texture orientations differ, we should be flipping images
when uploading them as textures, which we are doing already in the Texture
classes. And since we're doing that, and the UV conventions match, we should
_not_ need to flip the V coordinate.

This fixes texture orientation when previewing textures in Gaffer using
glsl/Texture.frag.