ObjectWriter : Remove dependency on VisibleRenderable
This does mean that we no longer write the bound into the header, but do we care any more? This dates back to the olden days when we'd render from a collection of .cob files, whereas nowadays we'd use a .scc or .abc file, both of which can embed bounds.