IECoreImage : Remove CurveTracer
This represents a slight problem for modularisation because it uses CurvesPrimitives, which are due to move to IECoreScene, and we don't want to have IECoreImage require IECoreScene. If we wanted to keep it, the plan would be to reimplement it as a simple ImageAlgo function, and have it return something simpler than a CurvesPrimitive - probably just V3fVectorData and IntVectorData. Since we currently have no need for it though, it's easier just to remove it for now.