DataTraits : Remove reference to SmoothSkinningData
I don't think there is any need for this, and we need to break the link between IECore and the upcoming IECoreScene. If we do need it, we can always inject it from IECoreScene.__init__, but really I think we should remove DataTraits entirely, and instead mirror the C++ API better by having nested members like `IntData.ValueType` etc.