IECoreScene : Move C++ tests from IECore
Rather than set up another C++ target using boost::unit_test, we instead just bind the tests as python functions so we can run them in the usual way with all the other tests. The boost framework has been a pain, and binding the C++ tests to python has worked out well in Gaffer.

At some point I intend to update Cortex to take the Gaffer approach, so that the tests are actual python modules that can be used at any time, and all the C++ tests are bound, with no dependency on boost::unit_test.