Remove Imath bindings
At the time these were added to Cortex, there were no official bindings for Imath. But now there are well established standard bindings, and the IECore module is imcompatible with them because it binds the same classes. We will use the standard bindings from now on.

I've kept the overloads for IECorePython::str and IECorePython::repr for now (moving them into IECore.cpp), as they are used by the Data bindings.