SimpleTypedDataBinding : Reorder bindings to fix Color3fData extraction
We use TypedDataFromType to register converters from T to TypedData<T>, and the order of registration affects the order in which boost::python tries the converters. The Cortex bindings for Imath::Color3 didn't declare Imath::V3f as a base class, but the PyImath bindings do. That means we need to register the color converters first, so that Color3f is converted to Color3fData by preference, and not V3fData.

This fixes a test failure in test/IECoreGL/ShadingTest.py, whereby testCsParameterTrumpsColorAttribute was declaring a "color" attribute using "Color3f", but it was appearing as V3fData on the C++ side.