SConstruct : Fix IECoreUSD build problems
Problem one was a linking error on Mac :

```
Undefined symbols for architecture x86_64:
  "__Py_NoneStruct", referenced from:
      __GLOBAL__sub_I_USDScene.cpp in USDScene.os
  "boost::python::converter::registry::lookup(boost::python::type_info)", referenced from:
      ___cxx_global_var_init.31 in USDScene.os
      ___cxx_global_var_init.32 in USDScene.os
```

It seems that USD doesn't cleanly separate its python bindings from the C++ libraries, so merely calling C++ code will force us to reference symbols provided by libpython. Building the usdEnv on top of the corePythonEnv ensures we link correctly to satisfy this unwanted dependency.

Problem two was that PYTHON_INCLUDE_PATH might be empty, in which case our CXXFLAGS contained an empty string, which broke the configuration stage. What we really wanted to use was PYTHON_INCLUDE_FLAGS, which is happily provided by the corePythonEnv anyway.

Also removed a duplicate inclusion of USD_INCLUDE_PATH in the CXXFLAGS>