SConstruct : Fix IECoreUSD configuration bug
The `corePythonEnv` includes libIECore in its LIBS variable. This means that the configuration check for the USD library used an environment which linked to IECore. This in turn meant that SCons was building the entirety of IECore _during the configure phase_. The `pythonEnv` doesn't include IECore in LIBS, so is save to use, and we were already manually adding libIECore after the configuration check had completed anyway.