IndexedIO : Move private factory out of header
The presence of the inline `getCreateFns()` method in the header was triggering some utterly inexplicable behaviour from Clang when compiling CompoundDataBase.cpp. The mere existence of `getCreateFns()` would cause Clang to emit the template instantiation for `TypedData<CompoundDataMap>` as private symbols rather as public ones, leading to linking errors down the line.