PerlinNoiseBinding : Fix symbol visibility problem on OSX
The issue here appears to be the same as the one tackled in the previous IndexedIO commit. If Clang sees VectorTypedData.h first, it is perfectly happy to consider all the VectorTypedData classes to have public visibility as we've specified. But if gets a sniff of any inline code involving `std::vector` _before_ it sees VectorTypedData.h, it changes its mind and hides all the VectorTypedData symbols.