TypedData classes : Declare as `extern template`
We already perform explicit template instantiations in the .cpp files, and prevent other translation units from instantiating the templates implicitly by hiding all the template implementations in .inl files that are not included by any headers. But even this is not enough! Compilers need to output typeinfo symbols for all classes, and because these are entirely implicitly generated rather than generated from source, the compiler can output them without seeing the .inl files. The C++11 `extern template` feature allows us to say "hey, we're going to instantiate this one ourselves", so that no symbols are generated for the template except at the point of explicit instantiation.

This fixes test failures on OSX, caused by multiple typeinfo symbols being exported for the TypedData classes, which in turn leads to boost::python not being able to detect the type of arguments correctly at runtime.