IECoreAlembic : Don't link against hdf5_hl
HDF5 1.8.11 won't build with GCC 6, and the latest 1.8.20 that does now refuses to build hdf5_hl when the thread safety option is turned on. Alembic _needs_ the thread safety option, so was updated so as not to need hdf5_hl. See https://github.com/alembic/alembic/issues/61 for more details.