SearchPath : Add platform-agnostic constructor
Also deprecate all methods which explicitly take a separator. These are just a cross-platform-compatibility trap, since the only sensible thing to pass on Linux and MacOS is ':' and the only sensible thing to pass on Windows is ';'. I can't find any uses of the existing constructors which use any separator other than ':', and ironically pretty much the only use for `getPaths()` I can find is `paths = sp.getPaths( ":" ).split( ":" )`, which is a just a very roundabout equivalent of `sp.paths`.

Any more esoteric needs that arise in the future can still be dealt with by manipulating the `SearchPath::paths` member directly.