ConfigLoader : Make `contextDict` argument optional
Neither Cortex nor Gaffer want to use this argument, but to date they have needed to pass `{}`. The big downside of this is that because the contextDict is shared between all config files, dependencies can creep in between one file and the next, leading to problems down the line. See https://github.com/GafferHQ/gaffer/pull/2457#discussion_r169282855 for one such example. When not passing a `contextDict`, a fresh dictionary is used for each and every config file, preventing this issue.