MeshPrimitive::createSphere : Round the number of segments
Previously we were flooring the number of segments, which created precision problems right around whole numbers of segments.

I also moved the +1 from inside the multiply to outside - it makes more sense for a half sphere to have half the number of segments, rather than half the number of vertices ( the difference would be visible as a jump when thetaMax = 179.99 vs 180 ).