ConfigLoader : Isolate config files from one another
Previously we were reusing the `contextDict` from one file to another. This makes it far too easy for one config file to inadvertently depend on the side effects of another, or worse, to stomp over the local variables from another. For an example of just how bad this can be, try putting this in a GUI startup file for Gaffer :

```
config = None
```

That blows away a variable needed by the colorspace config, and totally breaks the viewer.

In 03cdeeaf78b227802aa4dca988cc04a708fbe4a0 I fixed this for the cases where no `contextDict` was specified, but because `Gaffer.Application` needs to provide a context, I now need a fix for all cases.

I've deliberately broken backwards compatibility here to provoke discussion. I contend that the old approach was ridiculously fragile and needed to be fixed. The prior argument for not breaking compatibility was that internal IE code is relying on the old behaviour, but after a bit of searching through the codebase, I can't see where.