ObjectInterpolator : Use `DataAlgo::dispatch()`
This isn't quite as pleasing as the previous substitutions of `despatchTypedData()` :

- We've dropped the ReturnType typedef, which is nice
- The fall-through behaviour is more explicit, which is also nice
- The call to `dispatch()` is simpler
- But we've had to use enable_if to limit the types our payload function is applied to

I think this is the sticking point in deciding if this is a worthy successor to `despatchTypedData()`. The bright side is that with C++17 we'll be able to do the `enable_if` type stuff much more readably using `if constexpr` in a generic lambda. I think that might justify this approach, but I'm interested in other opinions.