MeshAlgoDistortions : Refactor internals into simpler form
This code was originally derived from an Op which used `despatchTypedData()` in an unnecessary attempt to support V3dVectorData P. The legacy of this was still visible in the packaging of the core logic into a functor. We can simplify everything by replacing the functor with a regular function call. Note that this is just moving code around - nothing has changed in the logic.