IECore : Don't modify `sys` module's dlopenflags
We needed this in the past because we were not managing symbol visibility correctly, so each module had different typeinfo symbols for the same type (for template types, on linux). We now only emit the typeinfo in the module that "owns" the type, so no longer need RTLD_GLOBAL.

This change was motivated by an obscure crash when generating the Gaffer documentation with the Mesa drivers in a Docker container. As far as I can tell, RTLD_GLOBAL was causing conflicts between the LLVM in the Mesa driver and the LLVM in Arnold, which caused crashes at shutdown.