SConstruct : Appleseed tests depend on Appleseed lib
A couple of us have now stared at this a bit, and we can't figure out why this wasn't already working.  Without this change, if the Appleseed lib is modified, running testAppleseed does trigger a rebuild, but then it refuses to run the actual test, saying it's up to date.  If anyone understands Scons, they could probably find a better solution to this.