ShaderNetwork : Make contained shaders immutable
This means that our lazily computed hash is no longer vulnerable to unknown changes made to shaders by external code. It also means that we can optimise `ShaderNetwork::copy()` by using shallow copies. The downside is forcing clients to use a get/modify/set sequence if they wish to modify a shader, and the client needing to know a bit about move semantics if they wish to avoid copies when adding shaders.