Canceller binding : Fix Python->C++ conversion for `IECore.Cancelled`
This was broken by 0c199d5543041b7be0ff24a052ce2e753c02d10c. `ExceptionAlgo::translatePythonException()` relies on `extract<Cancelled>()` to determine if the exception is an instance of `Cancelled`. When we used `boost::class` to bind Cancelled this worked automatically, but now we bind manually via `PyErr_NewException()` we are responsible for providing a Python->C++ type conversion ourselves.

Back in #773 when I introduced the Canceller, I was aware of the importance of this translation, but didn't include a test case, claiming it wasn't possible to test from a pure Python unit test. I consoled myself that I had tests in Gaffer that would catch the problem instead. That's true, but today they only caught the problem after a lengthy update cycle of GafferHQ/dependencies, leaving me back at square one. The lesson : put your tests as close to the thing they are testing as possible. This time I've managed to find a way of testing it, using the exception translation in the LRUCache bindings.