FromHoudiniGeomtryConverter : Remove AttributeRemap & RemapInfo.
The parallel attrib->primvar conversion caused the AttributeRemap tests to fail (because we need to know how many primvars will exist ahead of time). Rather than fix it, I decided to remove the feature, as it was fairly RenderMan specific, hasn't been used since IECoreHoudini was in its infancy (as far as I'm aware), and these days it'd be more appropriate to do those sorts of conversions either live in Houdini or as a post process using a tool like Gaffer.