OpenImageIOAlgo : Fix DataView copy constructor and copy assignment
When viewing StringData or StringVectorData, we populate `m_charPointers` to provide data in the layout required by OIIO. When copying, we need to ensure that the copy points to the data owned by its own `m_charPointers`, and not the source's. This bug was detected by Asan when running the GafferOSL tests.