TBBBinding : Bind tbb::global_control
In Gaffer we have been using the already-bound task_scheduler_init to limit the number of threads for the application, but that is not sufficient. If a thread creates an arena, then this ignores the task_scheduler_init and maxes out the thread count again. The global_control is the one true way to specify the maximum number of threads TBB may use.