Houdini LiveScene : Fix bug with mesh interpolation.
The previous approach worked for the simple case of one SOP containing one mesh (as per the existing test in FromHoudiniPolygonsConverter), but was incorrectly giving all meshes the same interpolation in the more realistic production case of one SOP containing many meshes. The new test exposes the bug and the code changes here fix it.