OpenImageIOAlgo : Support Rational type in image metadata
Some of ImageSequences that we received from the clients have the type
of the framesPerSecond value, in image metadata, set to Rational. At
read time this value is interpreted as v2i, which is absolutely fine
for future in memory manipulation. However when the image is written to
the filesystem it do not preserve or set the type to Ration hence the
further usage of this values by other applications may result in weird
behaviors, like for example playing the images sequences at 24000 FPS.

Example of proper set type, as printed by exrheader tool:
```
framesPerSecond (type rational): 24000/1001 (23.976)
```

Example of wrong (or not expected) type:
```
framesPerSecond (type v2i): (24000 1001)
```