SConstruct: We are moving away from using SideFx Alembic.
In Houdini 17.5, SideFx is shipping an openEXR lib with a custom suffix but
no python binding.
Since cortex needs the imath binding, we decided not to use their openEXR and
instead use ours.
The consequence of that is we can't use their Alembic either and thus we need
to compile IECoreAlembic with out headers and link against our library.
As a consequence, we need to preprend Houdini's header instead of appending them.