TextureLoader : support maximum resolution for loaded textures.
Will load the respective miplevel instead of the entire texture if the
resolution needs to be restricted. The texture cache is adjusted accordingly so
that we can cache textures at different resolutions.

BREAKING CHANGES:
`TextureLoader::load()` will now take an additional argument that specifies the
maximum resolution the texture can have.