SConstruct : Fix Mac build break
This fixes Mac build errors when using the `PYTHON_CONFIG` build variable rather than the `PYTHON_INCLUDE_PATH` and `PYTHON_LINK_FLAGS`. In this case, we don't include Python with `-isystem` and the configure step fails with the following error :

```
/Users/john/dev/gafferBuild/dependencies-0.54.2.0-source/gafferDependenciesBuild/lib/Python.framework/Versions/2.7/include/python2.7/unicodeobject.h:534:5: error: 'register' storage class specifier is deprecated and incompatible with C++17 [-Werror,-Wdeprecated-register]
    register PyObject *obj,     /* Object */
    ^~~~~~~~~
```

I don't know why a `C++17` error is triggered when we're explicitly passing `-std=c++11`, but it is. By testing for `boost/python.hpp` instead, we benefit from various magic boost does to suppress errors from the `Python.h` that it includes indirectly.