SConstruct : Only apply libraryPathEnvVar to non-Windows builds
Windows does not have a default library path. Normally library paths would go in PATH (and the rest of the Cortex build is set up that way) but because *Env["ENV"][libraryPathVar] assigns rather than appends to the library path, it will remove all the important PATH entries. So we just exclude this assignment on Windows and let OpenImageIO find the needed libraries using PATH.