IECoreMaya : Future-proof plugin symbol fetching
Note we're not using MAYA_API_VERSION in the actual symbol name, because it appears Maya doesn't use it. In 2019.2, MAYA_API_VERSION=20190200 but the expected symbol is 20190000. Hopefully this notion holds for 2020, but it remains to be seen...

Note also despite being documented in 2019 as existing since the beginning of Maya, MAYA_APP_VERSION does not actually exist prior to 2019...