Appleseed driver : Lift display initialisation out of tile loop
Really, it's a bit awkward that we're responsible for calling `DisplayLayer::initDisplay()` at all. It'd be  better dealt with in the DisplayLayer constructor, but we don't have a `Frame` available at that point. One option would be to pass a `Frame` to `highlightRegion()` and `writeTile()` and then initialise lazily in those, but for now I've followed the logic of the original code.