IECoreMaya SceneShape VP2 : Make unique MRenderItems per location.
We were previously reusing MRenderItems with matching topology, but we are not allowed to do this. MRenderItems are used for selection in VP2, to identify unique components. By reusing them, we were invalidating the component selection, effectively re-directing any selection back to one master component for each unique primitive.

We may have many more MRenderItems now, but selection is working, and we're still sharing the actual geometry (MVertexBuffers and MIndexBuffers), so hopefully this won't be a memory/performance concern.