IECoreMaya SceneShape VP2 : Combined the 3 internal caches
The primary motivation here was to use the single user-facing memory setting `IECORE_MAYA_VP2_MEMORY` directly as the memory limit of our internal cache. Previously we were arbitrarily splitting that memory budget between 3 caches.

A side benefit is that that we've decoupled the various primvar data (P, N, uv) as well as the wireframe vs non-wireframe indices for meshes, so we can compute them all independantly on-demand, and avoid the performance & memory cost if we don't need one of them. For example, if the user never activates Textured mode in the viewport, then we never compute the expandulated UVs.

I've also fixed the index buffers for PointsPrimitives, so they're now drawing correctly.