IECoreMaya : Fixed failing unit test related to FromMayaEnumPlugConverter not registering
Added a default constructor to FromMayaPlugConverter::Description which allows a converter
to be registered with the RTTI system without registering it with the factory mechanism.

The factory logic for enum plugs is already handled directly in FromMayaPlugConverter. We
just needed to register the type, which wasn't previously done due to the lack of a valid
registration function. Creating a new registration function just for enums is a bit overkill
since it would just be registering a short with conversions to shortData and stringData.