IECoreMaya : Reset SceneShape visibility when expanding SceneShapes with tags.
When new scene shapes are created, their visibility is set according to the draw tags.
However, if a new draw filter is set, and then a scene shape is re-expanded, it will not
re-display SceneShapes which were previously hidden. It will, however, hide SceneShape
which are no longer match the draw flags.

This change makes the visbility all expanded nodes respect the visibility of the new draw
tags.