Enum: Removed assertion in favor of `enum.Enum` behaviour
Since we moved to Python 3 compatible code, the assertion previously
used in the comparisons would break certain functionality, since it
would only now start to fail. Keeping it around breaks to many things
immediately. As a countermeasure we decided to let the comparison
return `False` instead. This helps in cases where we
are testing for variables memberships that would currently fail because
any comparison against anything else but a matching Enum class would
throw an assertion error, breaking some Gaffer functionality.

The downside here is that there rare case where two Enums that are not
of the same class, but were expected to match, would fail now silently.
Catching these cases with warning is hard, since we cannot determine
what and what is not a valid comparison without analyzing the code.