IECoreMaya SceneShape VP2 : Enable marquee selection for components
Note there is still an edge case that is not working: marquee selection in toggle mode (holding shift) does not work if you have child bounds drawing and geometry drawing both enabled, and the marquee hits both the child bounding box and the actual geometry. In this case we get a double selection reported as we have two active `SceneShapeSubSceneOverride::ComponentConverter`s and toggling flips the state twice, so we end up with the same selection we started with.