IECoreMaya : Added `toMayaAttributeName`, `fromMayaAttributeName`, and `visibilityOverrideName` to LiveScene
The added functions are convenience functions that will translate the cortex attribute
name to the maya attribute name. This was added to avoid client code from defining the
mapping in many different places.

`visibilityOverrideName` is the maya attribute name which should override IECoreScene.SceneInterface.visibilityName
This is defined so that client code does not hard code the attribute name. It's currently set to "ieVisibility".