UnicodeToStringBinding : Disable for Python 3 builds
In Python 3, boost::python automatically converts Python (unicode) strings to and from UTF-8 encoded `std::strings`, so we have better unicode support out of the box than what our custom `ExtractStringFromUnicode` was providing in Python 2.

It might be nice to improve `ExtractStringFromUnicode` to improve the Python 2 behaviour to match Python 3. The Python->C++ conversion could be dealt with easily by using `PyUnicode_AsUTF8String()` instead of `PyUnicode_AsASCIIString()` in `ExtractStringFromUnicode::construct()`. The C++->Python conversion might be trickier though. The default behaviour is to return a UTF8 encoded `str` object to Python, but we might prefer to return `unicode`. Either way, it's not possible to round-trip either `str` or `unicode` through C++ and back while maintaining the Python type, because we've lost the type information on conversion to `std::string`. For now it seems preferable to leave the Python 2 behaviour unchanged.