VectorTypedDataBinding : Add Python 3 comparison operators
Python 3 decided that the elegance of a single `__cmp__()` method that implemented all comparisons in a totally self-consistent way wasn't enough, and what was really needed was four different methods with the possibility of lots of internal contradiction. Ironically, C++ 20 is going in the opposite direction, with the `<=>` "spaceship" operator.