HexConversionTest : Limit range used to test `hexToDecChar()`
This is sufficient to get the tests passing with Python 3, but I think the idea of binding and testing `hexToDec<char>` and `decToHex<char>` as we are is fundamentally flawed. On the C++ side, a char is signed with a range of -127 to 127. But Boost::Python doesn't bind it as if it were an integer, it performs conversions to and from strings instead. This makes the negative part of the range unusable. The test was testing the range 0-255, which isn't representable as a char, and was therefore wrapping to negative numbers, and throwing an error when trying to convert a negative char to Python on the way out of `hexToDecChar()`.

I'm not away of any uses of this method in the wild, so for now I'm not putting any further effort into rationalising this. It's just one more thing I would delete given the chance.