Maya SceneShape VP2 : Defer componentIndex lookup until needed
We don't need to lookup the index unless we're in component mode. Its expensive because (a) it currently requires an extra IECoreGL render and (b) its a map search per location.

Note if we are in component mode, this commit actually increases the number of calls to selectionIndex(), but that will be addressed in a subsequent commit.