CachedReaderTest : Fix for IECoreScene split and Python 3
The test was trying to load `polySphereQuads.cob`, but that failed because that class is in IECoreScene now, so we weren't actually catching the exception from `PostProcessor` that we thought we were. Storing exceptions rather than using `assertRaisesRegex()` was also causing problems in Python 3, with a later failure in `test/IECore/WrapperGarbageCollector.py` caused by the PostProcessor being kept alive inside the exception object.